Made interfaces for structure reading and writing, Gson object with custom type adapters now used for both serialization and deserialization.
