+++ demo/src/main/resources/testWorld.json
@@ -14,5 +14,5 @@
       "data": "{name1:123,name2:\"sometext1\",name3:{subname1:456,subname2:\"sometext2\"}}"
     }
   ],
-  "blocks": "0,256;1,5;3,11;2,1;0,238;-1,3585"
+  "blocks": "0,256;1,5;3,11;1,1;0,238;-1,3585"
 }
