+++ src/main/java/com/mattworzala/canary/server/givemeahome/StructureReader.java
@@ -1,0 +1,7 @@
+package com.mattworzala.canary.server.givemeahome;
+
+import java.nio.file.Path;
+
+public interface StructureReader {
+    Structure readStructure(Path p);
+}
